Push *Corrected
Project Part 2:
- Adding support for FTP
- Client/Server both understand the client submitted command and
respond correctly

Current problem: data being sent to or from server is not being written
to the actual file that it is meant to be stored. Problem with the data
types of string vs. chars.

Problems with using File.open(nameOfFile), .open() does not accept
strings which are what the variables are currently.